requirements.txt
numpy
pandas
scikit-learn
opencv-python
flask
cassandradb
cassandra-driver
keras 
tensorflow
